chore: Restructure commands
